First Commit
Copy of Sigma16 for remote storing with changes to assembler line interpretation method so that the whitespace at the start of the line isn't required. However for labels to correctly work, a semi-colon is needed before them to denote them. Similarly, double semi-colons are required to denote comments. Slight change in how the base language is written but for fixing the requirement for whitespace at the start of all non-labelled lines, it is arguably worth it